samples/api_calls_cache/function/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore>=1.9.0
flask>=1.0.2